+++ src/modules/statetocolumn_mod.f90
@@ -113,10 +113,103 @@ module stateToColumn_mod
   logical :: slantPath_RA_nl
   logical :: calcHeightPressIncrOnColumn
   logical :: useFootprintForTovs
+  logical :: rejectObsNonMonotonicPressure
 
   integer, external :: get_max_rss
 
 contains 
+
+
+  !---------------------------------------------------------
+  ! pressureProfileMonotonicityCheck
+  !---------------------------------------------------------
+  subroutine pressureProfileMonotonicityCheck(obsSpaceData, column)
+    !
+    ! :Purpose: Check for non monotonic pressure profiles that can be computed in slantpathmode
+    !
+    implicit none
+
+    ! arguments
+    type(struct_obs), intent(inout)       :: obsSpaceData
+    type(struct_columnData), intent(inout):: column
+
+    ! locals
+    integer, parameter :: numWriteMax = 10
+    integer :: headerIndex, bodyIndex, iterationCount, singularIndex, levelIndex
+    integer :: pressureVarIndex
+    integer :: nlv
+    integer :: numWrites
+    real(8), pointer :: pressureProfile(:)
+    logical :: monotonicProfile
+    integer, parameter :: nPressureVar =2
+    character(len=4), parameter :: pressureVarList(nPressureVar)=['P_T ', 'P_M ']
+
+    write(*,*) ' '
+    write(*,*) 'pressureProfileMonotonicityCheck: START'
+    write(*,*) ' '
+    write(*,*) 'Memory Used: ',get_max_rss()/1024,'Mb'
+
+    numWrites = 0
+
+    call obs_set_current_header_list(obsSpaceData, 'TO')
+    HEADER: do
+      headerIndex = obs_getHeaderIndex(obsSpaceData)
+      if (headerIndex < 0) exit HEADER
+      do pressureVarIndex = 1, nPressureVar
+        pressureProfile => col_getColumn(column, headerIndex, pressureVarList(pressureVarIndex))
+        nlv = size(pressureProfile)
+        monotonicProfile = .true.
+        iterationCount = 0
+        iterationLoop:do
+          singularIndex = -1
+          levelSearch:do levelIndex = 1, nlv - 1
+            if ( pressureProfile(levelIndex) > pressureProfile(levelIndex+1)) then
+              singularIndex = levelIndex
+              exit levelSearch
+            end if
+          end do levelSearch
+          if ( singularIndex == -1 ) exit iterationLoop !regular profile or correction OK
+          iterationCount = iterationCount + 1
+          if (iterationCount == 1) then
+            monotonicProfile = .false.
+            if (numWrites < numWriteMax) then
+              numWrites = numWrites + 1
+              write(*,*) 'pressureProfileMonotonicityCheck: found non monotonic pressure profile:', &
+                   pressureVarList(pressureVarIndex), pressureProfile
+            end if
+          end if
+          if (singularIndex == 1) then !should never happen
+            write(*,*) 'pressureProfileMonotonicityCheck: ', pressureProfile(1:2)
+            call utl_abort('pressureProfileMonotonicityCheck: profile in the wrong order ?' &
+                 // pressureVarList(pressureVarIndex))
+          end if
+          pressureProfile(singularIndex) = 0.5d0 * ( pressureProfile(singularIndex - 1) + pressureProfile(singularIndex + 1) )
+          write(*,*) 'pressureProfileMonotonicityCheck: profile iteration', &
+               pressureVarList(pressureVarIndex), iterationCount
+        end do iterationLoop
+
+        ! if requested reject the corrected profile
+        if (.not. monotonicProfile .and. rejectObsNonMonotonicPressure) then
+          call obs_headSet_i(obsSpaceData,OBS_ST1,headerIndex, &
+               ibset( obs_headElem_i(obsSpaceData,OBS_ST1,headerIndex), 05))
+          call obs_set_current_body_list(obsSpaceData, headerIndex)
+          BODY: do
+            bodyIndex = obs_getBodyIndex(obsSpaceData)
+            if (bodyIndex < 0) exit BODY
+            if (rejectObsNonMonotonicPressure) then
+              call obs_bodySet_i(obsSpaceData, OBS_ASS, bodyIndex, obs_notAssimilated)
+              call obs_bodySet_i(obsSpaceData, OBS_FLG, bodyIndex, &
+                   ibset(obs_bodyElem_i(obsSpaceData, OBS_FLG, bodyIndex),9))
+            end if
+          end do BODY
+        end if
+      end do ! loop on pressure variables
+     
+    end do HEADER
+
+    write(*,*) 'pressureProfileMonotonicityCheck: END'
+
+  end subroutine pressureProfileMonotonicityCheck
 
   !---------------------------------------------------------
   ! latlonChecksAnlGrid
@@ -306,7 +399,7 @@ contains
     type(kdtree2), pointer  :: tree
 
     namelist /nams2c/ slantPath_TO_nl, slantPath_TO_tlad, slantPath_RO_nl, slantPath_RA_nl, calcHeightPressIncrOnColumn
-    namelist /nams2c/ useFootprintForTovs 
+    namelist /nams2c/ useFootprintForTovs, rejectObsNonMonotonicPressure 
 
     write(*,*) 's2c_setupInterpInfo: STARTING'
     write(*,*) 'Memory Used: ',get_max_rss()/1024,'Mb'
@@ -329,6 +422,7 @@ contains
       slantPath_RA_nl   = .false.
       calcHeightPressIncrOnColumn = .false.
       useFootprintForTovs = .false.
+      rejectObsNonMonotonicPressure =.true.
 
       if ( .not. utl_isNamelistPresent('NAMS2C','./flnml') ) then
         if ( mpi_myid == 0 ) then
@@ -1322,6 +1416,9 @@ contains
     if (calcHeightPressIncrOnColumn) then
       call gsv_deallocate( stateVector )
     end if
+    
+    if (slantPath_TO_tlad) call pressureProfileMonotonicityCheck(obsSpaceData, columnTrlOnAnlIncLev)
+
     call tmg_stop(167)
 
   end subroutine s2c_tl
@@ -1537,6 +1634,8 @@ contains
     end if
 
     call gsv_deallocate( statevector_VarsLevs )
+
+    if (slantPath_TO_tlad) call pressureProfileMonotonicityCheck(obsSpaceData, columnTrlOnAnlIncLev)
 
     call tmg_stop(168)
 
@@ -1908,6 +2007,8 @@ contains
     end do OBSBATCH
 
     call gsv_deallocate( statevector_VarsLevs )
+
+    if (slantPath_TO_nl) call pressureProfileMonotonicityCheck(obsSpaceData, column)
 
     write(*,*) 's2c_nl: FINISHED'
     write(*,*) 'Memory Used: ',get_max_rss()/1024,'Mb'